done changes as per ui design
