+++ client/src/Components/Feature/Feature.css
@@ -1,0 +1,4 @@
+.bg-feature {
+    background-color: rgb(217, 213, 252);
+        color: black;
+}
