+++ client/src/Components/Hero/Hero.css
@@ -1,0 +1,18 @@
+.background {
+    background-color: rgba(77, 68, 181, 1);
+    color: white;
+}
+
+.button {
+    background-color: rgba(217, 217, 217, 1)  ;
+    color: black;
+    font-size: 18px;
+    font-weight: 400;
+}
+
+.button:hover {
+    background-color: rgba(193, 187, 235, 1);
+}
+
+
+
