+++ client/src/Components/NavBar/Header.css
@@ -1,0 +1,46 @@
+.navbar {
+    background-color: white;
+    padding: 20px;
+}
+
+.btn {
+    background-color: #4D44B5;
+    color: white;
+    margin-left: 30px;
+}
+
+.btn:hover {
+    background-color: rgba(193, 187, 235, 1);
+}
+
+.nav-link:hover {
+    color: #ED5D34;
+}
+
+.nav-link {
+    font-weight: 400;
+        font-size: 1.2rem;
+        padding-left: 70px;
+        height: 60px;
+}
+.nav-link {
+    padding-top: 20px;
+}   
+.logo {
+    padding-left: 50px;
+    padding-right: 70px;
+}
+.nav-item-left {
+    color: #4D44B5;
+    font-weight: 500;
+    font-size: 1.2rem;
+    padding-left: 70px;
+    height: 60px;
+    text-decoration: none;
+}
+.nav-item-left:hover {
+    color: #ED5D34;
+}
+.btn-pinks {
+    padding-left: 90px;
+}
