--- client/src/Hero/Hero.css
@@ -1,23 +1,0 @@
-.background {
-    background-color: rgba(77, 68, 181, 1);
-    color: white;
-}
-
-.name {
-    color: rgba(249, 201, 31, 1);
-}
-
-.button {
-    background-color: white;
-    color: black;
-    padding-top: 20px;
-}
-
-.button:hover {
-    background-color: rgba(193, 187, 235, 1);
-}
-.main {
-    color: white;
-}
-
-
