+++ package.json
@@ -75,6 +75,7 @@
     "winston": "^3.11.0"
   },
   "engines": {
+    "node": "^18.3 || ^20",
     "npm": ">=8"
   }
 }
